Revert "Type implements if it is triggered by an error state (default to false) and build payload for error state VS non-separately"
This reverts commit d2ad060e9be0c2c28ecca83c75ecda53cb654e40.

Nevermind, I actually want Metrics to be triggered based on the value being set. So only return EVENTS_PROCESSED if the survey was actually processed or EXCEPTION_COUNT if it did fail. Pushing more of this logic to the class returning a meaningful value rather than returning a default value that has to then be filtered arbitrarily.